refactor modal to use mount / unmount
